Crate flowclib

This is the rust flow Compiler Library. It can be linked with other code to produce a flow compiler, such as the flowc command line flow compiler.

Modules

compiler

reads [flow definitions][flowcore::model::flow_definition] into memory, flattens and connects that into a graph of [runtime functions][flowcore::model::runtime_function::RuntimeFunction] Compile module that loads flow definition, compiles flows and tables and then generates JSON manifest of processes

dumper

used to output a human readable version of the model and compiler tables to help debug compiler problems

errors

contains errors::Error that other modules in this crate will use errors::*; to get access to everything error_chain creates.

generator

takes care of generating the [flow’s manifest][flowcore::model::flow_manifest::FlowManifest] (for later execution by a flow runner) from the compiler tables

info

provides methods to get information about this version of the flowclib library Info module provides methods to get additional information about the flowclib library