Module flowclib::compiler::loader

Traits

Deserializer
All deserializers have to implement this trait for content deserialization, plus a method to return their name to be able to inform the user of which deserializer was used

Validate
Many structs in the model implement the Validate method which is used to check the description deserialized from file obeys some additional constraints that cannot be expressed in the struct definition in serde

Functions

load
Load a context process definition from url, recursively loading all sub-processes referenced.

load_metadata
load library metadata from the given url using the provider. Currently it used the package table of Cargo.toml as a source but it could easily use another file as along as it has the required fields to satisfy MetaData struct