Function flowclib::dumper::dump::dump_tables

pub fn dump_tables(tables: &CompilerTables, output_dir: &Path) -> Result<()>

Dump the compiler tables of a loaded flow in human readable format to a specified output directory.

Example

use std::env;
use url::Url;
use flowcore::provider::Provider;
use flowcore::meta_provider::MetaProvider;
use flowcore::errors::Result;
use flowcore::model::process::Process::FlowProcess;
use std::collections::{BTreeSet, HashSet};
use simpath::Simpath;
use std::path::PathBuf;
use tempdir::TempDir;

let lib_search_path = Simpath::new("FLOW_LIB_PATH");
let provider = MetaProvider::new(lib_search_path, PathBuf::from("/"));

let mut url = Url::from_file_path(env::current_dir().unwrap()).unwrap();
url = url.join("samples/hello-world/root.toml").unwrap();

let mut source_urls = BTreeSet::<(Url, Url)>::new();
let output_dir = TempDir::new("flow-test").expect("A temp dir").into_path();

if let Ok(FlowProcess(mut flow)) = flowclib::compiler::parser::parse(&url,
                                                          &provider,
                                                          &mut source_urls) {
    let tables = flowclib::compiler::compile::compile(&mut flow,
                                                      &output_dir, true, true,
                                                      #[cfg(feature = "debugger")] &mut source_urls
                                                      ).unwrap();

    flowclib::dumper::dump::dump_tables(&tables, &output_dir).unwrap();
}