Function flowclib::dumper::dump_flow::dump_flow

pub fn dump_flow(
    flow: &Flow, 
    output_dir: &Path, 
    provider: &dyn LibProvider, 
    dump_files: bool, 
    dot_files: bool
) -> Result<()>

Dump a human readable representation of loaded flow definition (in a Flow structure) to a file in the specified output directory

Example

use std::env;
use url::Url;
use flowcore::lib_provider::{LibProvider, MetaProvider};
use flowcore::errors::Result;
use flowclib::model::process::Process::FlowProcess;
use tempdir::TempDir;
use std::collections::HashSet;
use simpath::Simpath;

let lib_search_path = Simpath::new("FLOW_LIB_PATH");
let provider = MetaProvider::new(lib_search_path);

let mut url = url::Url::from_file_path(env::current_dir().unwrap()).unwrap();
url = url.join("samples/hello-world/context.toml").unwrap();

let mut source_urls = HashSet::<(Url, Url)>::new();
if let Ok(FlowProcess(mut flow)) = flowclib::compiler::loader::load(&url,
                                                   &provider,
                                                   &mut source_urls) {

    // strip off filename so output_dir is where the context.toml file resides
    let output_dir = TempDir::new("flow").unwrap().into_path();

    // dump the flows compiler data and dot graph into files alongside the 'context.toml'
    flowclib::dumper::dump_flow::dump_flow(&flow, &output_dir, &provider, true, true).unwrap();
}