Function flowclib::dumper::dump_tables::dump_tables

pub fn dump_tables(tables: &GenerationTables, output_dir: &Path) -> Result<()>

Dump the compiler tables of a loaded flow in human readable format to a specified output directory.

Example

use std::env;
use url::Url;
use provider::lib_provider::Provider;
use provider::errors::Result;
use flowclib::model::process::Process::FlowProcess;
use provider::lib_provider::LibProvider;

struct DummyProvider {}

impl LibProvider for DummyProvider {
    fn resolve_url(&self, url: &Url, default_filename: &str, _ext: &[&str]) -> Result<(Url, Option<String>)> {
        Ok((url.clone(), None))
    }

    fn get_contents(&self, _url: &Url) -> Result<Vec<u8>> {
        Ok("flow = \"dummy\"".as_bytes().to_owned())
    }
}

let dummy_provider = DummyProvider {};
let mut url = url::Url::from_file_path(env::current_dir().unwrap()).unwrap();
url = url.join("samples/hello-world-simple/context.toml").unwrap();

if let FlowProcess(mut flow) = flowclib::compiler::loader::load(&url,
                                                          &dummy_provider).unwrap() {
    let tables = flowclib::compiler::compile::compile(&mut flow).unwrap();
    let output_dir = tempdir::TempDir::new("flow").unwrap().into_path();

    let tables = flowclib::compiler::compile::compile(&mut flow).unwrap();
    let output_dir = tempdir::TempDir::new("dumper").unwrap().into_path();

    flowclib::dumper::dump_tables::dump_tables(&tables, &output_dir).unwrap();
}