Function flowclib::dumper::flow_to_dot::dump_flow

pub fn dump_flow(
    flow: &FlowDefinition,
    output_dir: &Path,
    provider: &dyn Provider
) -> Result<()>

Create a .dot format directed graph of a loaded flow definition

Example

use std::env;
use url::Url;
use flowcore::provider::Provider;
use flowcore::meta_provider::MetaProvider;
use flowcore::errors::Result;
use flowcore::model::process::Process::FlowProcess;
use tempdir::TempDir;
use std::collections::BTreeMap;
use simpath::Simpath;
use std::path::PathBuf;

let lib_search_path = Simpath::new("FLOW_LIB_PATH");
let provider = MetaProvider::new(lib_search_path, PathBuf::from("/"));

let mut url = Url::from_file_path(env::current_dir().unwrap()).unwrap();
url = url.join("samples/hello-world/root.toml").unwrap();

if let Ok(FlowProcess(mut flow)) = flowclib::compiler::parser::parse(&url, &provider) {
    // strip off filename so output_dir is where the root.toml file resides
    let output_dir = TempDir::new("flow").unwrap().into_path();

    // dump the flows compiler data and dot graph into files alongside the 'root.toml'
    flowclib::dumper::flow_to_dot::dump_flow(&flow, &output_dir, &provider).unwrap();
}