Struct flowclib::model::datatype::DataType

pub struct DataType(_);

Implementations

impl DataType

pub fn map<__SHRINKWRAP_T, __SHRINKWRAP_F: FnMut(String) -> __SHRINKWRAP_T>(
self,
f: __SHRINKWRAP_F
) -> __SHRINKWRAP_T

Map a function over the wrapped value, consuming it in the process.

pub fn map_ref<__SHRINKWRAP_T, __SHRINKWRAP_F: FnMut(&String) -> __SHRINKWRAP_T>(
&self,
f: __SHRINKWRAP_F
) -> __SHRINKWRAP_T

Map a function over the wrapped value without consuming it.

impl DataType

pub fn valid(&self) -> Result<()>

pub fn is_array(&self) -> bool

pub fn is_generic(&self) -> bool

pub fn array_of(&self, second: &Self) -> bool

Determine if this data type is an array of the other

pub fn within_array(&self) -> Result<Self>

Get the data type the array holds

pub fn type_string(value: &Value) -> String

Take a json data value and return the type string for it, recursively going down when the type is a container type (Array or Map(Object))

pub fn array_order(&self) -> Result<i32>

Take a string description of a DataType and determine how deeply nested in arrays it is