Struct flowclib::model::route::Route

pub struct Route(pub String);

Implementations

impl Route

pub fn map<__SHRINKWRAP_T, __SHRINKWRAP_F: FnMut(String) -> __SHRINKWRAP_T>(
self,
f: __SHRINKWRAP_F
) -> __SHRINKWRAP_T

Map a function over the wrapped value, consuming it in the process.

pub fn map_ref<__SHRINKWRAP_T, __SHRINKWRAP_F: FnMut(&String) -> __SHRINKWRAP_T>(
&self,
f: __SHRINKWRAP_F
) -> __SHRINKWRAP_T

Map a function over the wrapped value without consuming it.

pub fn map_mut<__SHRINKWRAP_T, __SHRINKWRAP_F>(
&mut self,
f: __SHRINKWRAP_F
) -> __SHRINKWRAP_T where
__SHRINKWRAP_F: FnMut(&mut String) -> __SHRINKWRAP_T,

Map a function over the wrapped value, potentially changing it in place.

impl Route

Route is used to locate Processes (Flows or Functions), their IOs and sub-elements of a data structure within the flow hierarchy

Examples "/my-flow" -> The flow called "my-flow, anchored at the root of the hierarchy, i.e. the context "/my-flow/sub-flow" -> A flow called "sub-flow" that is within "my-flow" "/my-flow/sub-flow/function" -> A function called "function" within "sub-flow" "/my-flow/sub-flow/function/input_1" -> An IO called "input_1" of "function" "/my-flow/sub-flow/function/input_1/1" -> An array element at index 1 of the Array output from "input_1" "/my-flow/sub-flow/function/input_2/part_a" -> A part of the Json structure output by "input_2" called "part_a"

pub fn sub_route_of(&self, other: &Route) -> Option<Route>

sub_route_of returns an Option indicating if self is a subroute of other (i.e. self is a longer route to an element under the other route) Return values None - self is not a sub-route of other (e.g. ("/my-route1", "/my-route2") (e.g. ("/my-route1", "/my-route1/something") Some(Route::from("")) - self and other are equal (e.g. ("/my-route1", "/my-route1") Some(Route::from(diff)) - self is a sub-route of other - with diff added (e.g. ("/my-route1/something", "/my-route1")

pub fn insert<R: AsRef<str>>(&mut self, sub_route: R) -> &Self

Insert another Route at the front of this Route

pub fn extend(&mut self, sub_route: &Route) -> &Self

Extend a Route by appending another Route to the end, adding the '/' separator if needed

pub fn route_type(&self) -> RouteType

pub fn pop(&self) -> (Cow<'_, Route>, Option<Route>)

Return a route that is one level up, such that /context/function/output/subroute -> /context/function/output

pub fn without_trailing_array_index(&self) -> (Cow<'_, Route>, usize, bool)

Return the io route without a trailing number (array index) and if it has one or not If the trailing number was present then return the route with a trailing '/'